requirements/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Django == 4.1
backports.zoneinfo == 0.2.1
django_ckeditor == 6.5.0
future == 0.18.2
ignition-api == 8.1.20.post3
jpype1 == 1.4.0
keyring == 18.0.1
mrlpy == 1.0.0
mypy-extensions == 0.4.3
openbaton-cli == 6.0.0
org-geppetto-recording == 0.1.0
py4grid == 1.0.5
python-plugin-sdk == 6.0.1
selector-standardization-beam == 0.4.15
selector-standardizers == 0.8.3
psycopg2-binary == 2.9.3
urllib3 == 1.25.8